requirements.txt
AfricastalkingGateway==1.8
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
requests==2.18.3
urllib3==1.22
Werkzeug==0.12.2
